Reduced the dataset obob_fasted_glc_lac_gln_ala_multiple_labels for tests that use it and moved it into data/tests/multiple_labels.
Removed 1 test.

Files checked in: DataRepo/tests/loading/test_load_accucor_msruns.py DataRepo/tests/test_models.py DataRepo/data/tests/multiple_labels/alafasted_cor.xlsx DataRepo/data/tests/multiple_labels/animal_sample_table.xlsx DataRepo/data/tests/multiple_labels/compounds.tsv DataRepo/data/tests/multiple_labels/glnfasted1_cor.xlsx DataRepo/data/tests/multiple_labels/loading.yaml DataRepo/data/tests/multiple_labels/protocols.tsv DataRepo/data/tests/multiple_labels/tissues.tsv
