+++ DataRepo/data/tests/multiple_labels/compounds.tsv
@@ -1,0 +1,15 @@
+Compound	Formula	HMDB ID	Synonyms
+3-hydroxybutyrate	C4H8O3	HMDB0000011	3HB;BHB;beta-hydryoxybutyric acid;beta-hydroxybutyrate;3-hydroxybutyric acid;3-HB;B-HB
+a-ketoglutarate	C5H6O5	HMDB0000208	2-ketoglutarate;2-ketoglutaric acid;alpha-ketoglutaric acid;a-ketoglutaric acid;oxoglutarate;2-oxoglutarate;2-oxoglutaric acid
+alanine	C3H7NO2	HMDB0000161	ala;L-alanine
+citrate	C6H8O7	HMDB0000094	citric acid
+glucose	C6H12O6	HMDB0000122	dextrose;Glucose;D-Glucose
+glucose-6-phosphate	C6H13O9P	HMDB0001401	GLC6p;G6P
+glutamate	C5H9NO4	HMDB0000148	glu;2-aminoglutarate;2-aminoglutaric acid
+glutamine	C5H10N2O3	HMDB0000641	gln
+isocitrate	C6H8O7	HMDB0000193	
+lactate	C3H6O3	HMDB0000190	lactic acid;2-hydroxypropanoate;2-hydroxypropanoic acid
+malate	C4H6O5	HMDB0000156	malic acid;hydroxysuccinate;hydroxysuccinic acid
+pyruvate	C3H4O3	HMDB0000243	pyruvic acid;2-ketopropionic acid;2-oxopropanoic acid;2-oxopropanoate;2-ketopropionate
+serine	C3H7NO3	HMDB0000187	ser
+succinate	C4H6O4	HMDB0000254	succinic acid
