+++ DataRepo/data/tests/multiple_labels/loading.yaml
@@ -1,0 +1,20 @@
+---
+protocols: "protocols.tsv"
+tissues: "tissues.tsv"
+compounds: "compounds.tsv"
+animals_samples_treatments:
+  table: "animal_sample_table.xlsx"
+  skip_researcher_check: true
+accucor_data:
+  accucor_files:
+    - name: "glnfasted1_cor.xlsx"
+      isocorr_format: true
+    - name: "alafasted_cor.xlsx"
+      isocorr_format: true
+  lc_protocol: "polar-HILIC-25-min"
+  instrument: "default instrument"
+  date: "2021-06-06"
+  researcher: "Xianfeng Zeng"
+  new_researcher: false
+  skip_samples:
+    - "bk"
