+++ DataRepo/data/tests/multiple_labels/tissues.tsv
@@ -1,0 +1,16 @@
+Name	Description
+brown_adipose_tissue	brown adipose tissue
+brain	whole brain
+gastrocnemius	gastrocnemius muscle
+white_adipose_tissue_gonadal	gonadal white adipose tissue
+heart	heart muscle (ventricle)
+kidney	kidney
+liver	liver
+lung	lung
+pancreas	pancreas
+quadricep	quadricep muscle
+small_intestine	small intestine
+soleus	soleus
+spleen	spleen
+serum_plasma_tail	serum or plasma collected from tail snip
+feces	feces
